+++ subsys/bluetooth/mesh/adv_legacy.c
@@ -137,8 +137,15 @@ static void adv_thread(void *p1, void *p2, void *p3)
 				 * to bt_mesh_adv_start:
 				 */
 				adv_timeout = SYS_FOREVER_MS;
-				bt_mesh_proxy_adv_start();
-				BT_DBG("Proxy Advertising");
+				if (bt_mesh_is_provisioned()) {
+					if (IS_ENABLED(CONFIG_BT_MESH_GATT_PROXY)) {
+						(void)bt_mesh_proxy_adv_start();
+						BT_DBG("Proxy Advertising");
+					}
+				} else if (IS_ENABLED(CONFIG_BT_MESH_PB_GATT)) {
+					(void)bt_mesh_prov_adv_start();
+					BT_DBG("PB-GATT Advertising");
+				}
 
 				buf = net_buf_get(&bt_mesh_adv_queue,
 						  SYS_TIMEOUT_MS(adv_timeout));
