+++ subsys/bluetooth/mesh/main.c
@@ -305,7 +305,7 @@ int bt_mesh_init(const struct bt_mesh_prov *prov,
 		return err;
 	}
 
-	if (IS_ENABLED(CONFIG_BT_MESH_GATT)) {
+	if (IS_ENABLED(CONFIG_BT_MESH_GATT_PROXY)) {
 		bt_mesh_proxy_init();
 	}
 
@@ -354,14 +354,13 @@ int bt_mesh_start(void)
 		bt_mesh_beacon_disable();
 	}
 
-	/* For PB-GATT provision, will enable in le disconnect handler. */
-	if (bt_mesh_prov_link.bearer->type == BT_MESH_PROV_ADV) {
+	if (!IS_ENABLED(CONFIG_BT_MESH_PROV) || !bt_mesh_prov_active() ||
+	    bt_mesh_prov_link.bearer->type == BT_MESH_PROV_ADV) {
 		if (IS_ENABLED(CONFIG_BT_MESH_PB_GATT)) {
 			(void)bt_mesh_proxy_prov_disable();
 		}
 
-		if (IS_ENABLED(CONFIG_BT_MESH_GATT_PROXY) &&
-		    bt_mesh_gatt_proxy_get() != BT_MESH_GATT_PROXY_NOT_SUPPORTED) {
+		if (IS_ENABLED(CONFIG_BT_MESH_GATT_PROXY)) {
 			(void)bt_mesh_proxy_gatt_enable();
 			bt_mesh_adv_update();
 		}
